refactor(api): extract precision boundary
